Testing GWT showed changes needed: v1.4.4.1
This release only affects GWT, since the only files that changed are the
.gwt.xml module file, the project file itself, and README.md . This is
necessary to use BlazingChain on GWT, however, and the <inherits> line
has changed since the previous one turned out not to work in practice.
The README has been updated with the new, only-slightly-changed
instructions, plus code usage.