Bugfix version 1.4.4.2; avoid corrupting large texts
Thanks to a bug reported via email, I was able to track down the issue
to some subtly incorrect code where `bits`, a char, was used like an int
and more than 16 bits are assigned to it (this truncates the upper
bits). This only happened if `dictionary`, an ArrayList of String
containing repeated segments of text to compress, exceeded a size of
65536, which never happened on the small sample data. The fix was simple
(use an int instead of reusing a char where a char isn't applicable),
and now version 1.4.4.2 is out. This bug doesn't appear to affect the
original lz-string library for JS by pieroxy, but probably could affect
lz-string4java by rufushuang , unless it's been fixed already there or
my changes introduced the bug.